Move ExitCodes out of `redbot.core.bot`
Avoids circular import
